pi-display/requirements.txt
Pillow>=9.0.0
requests>=2.28.0
pygame>=2.0.0
evdev>=1.6.0
# waveshare-epd  # Install from Waveshare repo for your specific display
